requirements.txt
pdfplumber
pytesseract
pdf2image
sentence-transformers
qdrant-client
langchain
openai
torch
Pillow
python-dotenv

fastapi
uvicorn
# python-dotenv
# sentence-transformers
# qdrant-client
openai
langchain-text-splitters
# pdfplumber
pydantic
